Chore: Removed metaclass and separated the pipeline logic so it can be tested separately.
