+++ pydatalab/src/pydatalab/pipeline_block/base.py
@@ -4,9 +4,6 @@ import warnings
 from pathlib import Path
 from typing import Any
 
-import bokeh.embed
-import pandas as pd
-
 from pydatalab import __version__
 from pydatalab.config import CONFIG
 from pydatalab.file_utils import get_file_info_by_id
@@ -18,13 +15,13 @@ from pydatalab.pipeline_block.block_stages import (
     PlotterStage,
     ProcessorStage,
 )
+from pydatalab.pipeline_block.pipeline import Pipeline
+from pydatalab.pipeline_block.utils import generate_random_id
 
 __all__ = ("PipelineDataBlock",)
 
-from pydatalab.pipeline_block.utils import generate_random_id
 
-
-class MetaPipelineDataBlock(type):
+class PipelineDataBlock:
     name: str = "base"
     """The human-readable block name specifying which technique
     or file format it pertains to.
@@ -45,154 +42,8 @@ class MetaPipelineDataBlock(type):
     supplied during block init.
     """
 
-    multi_file: bool = False
-    """Whether this block can accept multiple files as input."""
-
-    parser_functions: list[ParserStage]
-    """A list of methods that will parse files for this datablock."""
-
-    processor_functions: list[list[ProcessorStage]]
-    """A list of processor stages that will operate on the data for this datablock."""
-
-    plotter_function: PlotterStage
-    """ The plotter that will create the plot from the data in this datablock."""
-
-    event_functions: dict[str, EventStage]
-    """ The event stage functions, used when calling an event for a partial parameter update """
-
-    @staticmethod
-    def processor(df: pd.DataFrame) -> pd.DataFrame:
-        return df
-
-    @staticmethod
-    def plotter(df: pd.DataFrame) -> Any:
-        from pydatalab.bokeh_plots import selectable_axes_plot
-
-        if df is None:
-            return None
-        plot = selectable_axes_plot(
-            df,
-            plot_points=True,
-            plot_line=False,
-            show_table=True,
-        )
-        return bokeh.embed.json_item(plot)
-
-    @staticmethod
-    def null_event(data: dict, **kwargs):
-        """A null debug event that does nothing but logs its kwargs and overwrites the data dict with the args."""
-        LOGGER.debug(
-            "Null event received by pipeline data block %s with kwargs: %s",
-            data["blocktype"],
-            kwargs,
-        )
-        data["kwargs"] = kwargs["kwargs"]
-
-    def add_parser(self, parser_function: ParserStage | Any) -> None:
-        self.add_parsers([parser_function])
-
-    def add_parsers(self, parser_functions: list[ParserStage] | Any) -> None:
-        if self.parser_functions is None:
-            self.parser_functions = []
-        self.parser_functions.extend(parser_functions)
-
-    def add_processor(self, processor_function: ProcessorStage) -> None:
-        if self.processor_functions is None:
-            self.processor_functions = []
-        self.processor_functions.append([processor_function])
-
-    def add_stage_of_processors(self, processor_functions: list[ProcessorStage]) -> None:
-        if self.processor_functions is None:
-            self.processor_functions = []
-        self.processor_functions.append(processor_functions)
-
-    def set_plotter(self, plotter_function: PlotterStage) -> None:
-        self.plotter_function = plotter_function
-
-    def __init__(
-        cls,
-        name,
-        bases,
-        namespace,
-        *,
-        parser: ParserStage | list[ParserStage] | None = None,
-        processor: ProcessorStage | list[list[ProcessorStage]] | None = None,
-        plotter: PlotterStage | None = None,
-        events: dict[str, EventStage] | None = None,
-        blocktype: str | None = blocktype,
-        block_name: str | None = name,
-        description: str | None = description,
-        accepted_file_extensions: tuple[str, ...] | None = accepted_file_extensions,
-        multi_file: bool = multi_file,
-        defaults=None,
-    ):
-        super().__init__(cls)
-
-        # Set variables here
-        if defaults is None:
-            defaults = {}
-        cls.name = namespace.get("name", block_name)
-        cls.description = namespace.get("description", description)
-        cls.accepted_file_extensions = namespace.get(
-            "accepted_file_extensions", accepted_file_extensions
-        )
-        cls.blocktype = namespace.get("blocktype", blocktype)
-        cls.multi_file = namespace.get("multi_file", multi_file)
-        cls.defaults = namespace.get("defaults", defaults)
-
-        parser = namespace.get("parser", parser)
-        processor = namespace.get("processor", processor)
-        plotter = namespace.get("plotter", plotter)
-
-        # Set default default pipeline stages
-        if not parser:
-            # TODO default parser
-            parser = []
-            pass
-        if not processor:
-            processor = ProcessorStage(cls.processor, list_df_input=False)
-        if not plotter:
-            plotter = PlotterStage(cls.plotter)
-
-        cls.parser_functions = []
-        cls.processor_functions = []
-        cls.event_functions = {}
-
-        # Check types and assign to pipeline
-        if type(parser) is ParserStage:
-            parser = [parser]
-        if isinstance(processor, ProcessorStage):
-            cls.add_processor(processor)
-        else:
-            cls.processor_functions = processor
-        cls.add_parsers(parser)
-        cls.set_plotter(plotter)
-        if events:
-            cls.event_functions = events
-        if "null_event" not in cls.event_functions:
-            cls.event_functions["null_event"] = EventStage(cls.null_event)
-
-
-class PipelineDataBlock(metaclass=MetaPipelineDataBlock):
-    name: str = "base"
-    """The human-readable block name specifying which technique
-    or file format it pertains to.
-    """
-
-    blocktype: str = "generic pipeline"
-    """A short (unique) string key specifying the type of block."""
-
-    description: str = "Generic pipeline Block"
-    """A longer description outlining the purpose and capability
-    of the block."""
-
-    accepted_file_extensions: tuple[str, ...] | None = ()
-    """A list of file extensions that the block will attempt to read."""
-
-    defaults: dict[str, Any] = {}
-    """Any default values that should be set if they are not
-    supplied during block init.
-    """
+    pipeline: Pipeline = Pipeline()
+    """Pipeline to object to run."""
 
     multi_file: bool = False
     """Whether this block can accept multiple files as input."""
@@ -201,18 +52,6 @@ class PipelineDataBlock(metaclass=MetaPipelineDataBlock):
     """Base class for a data block. Has a metaclass of MetaPipelineDataBlock."""
     version: str = __version__
     """The implementation version of this particular block."""
-
-    parser_functions: list[ParserStage]
-    """A list of methods that will parse files for this datablock."""
-
-    processor_functions: list[list[ProcessorStage]]
-    """A list of processor stages that will operate on the data for this datablock."""
-
-    plotter_function: PlotterStage
-    """ The plotter that will create the plot from the data in this datablock."""
-
-    event_functions: dict[str, EventStage]
-    """ The event stage functions, used when calling an event for a partial parameter update """
 
     def __init_subclass__(cls, **kwargs):
         super().__init_subclass__()
@@ -266,6 +105,7 @@ class PipelineDataBlock(metaclass=MetaPipelineDataBlock):
             self.__class__.__name__,
             item_id,
         )
+        self.pipeline = self.__class__.pipeline.clone()
 
     def to_db(self) -> dict:
         """returns a dictionary with the data for this
@@ -283,11 +123,7 @@ class PipelineDataBlock(metaclass=MetaPipelineDataBlock):
             exclude_none=True,
         )
 
-    def perform_entire_pipeline(self):
-        """
-        Performs an entire complete pipeline with no caching or async operations.
-        Used for both testing and single threaded pipelines where caching is not an option.
-        """
+    def perform_operations(self):
         # First step - retrieve the file(s)
         if ("file_id" not in self.data and not self.multi_file) or (
             "file_ids" not in self.data and self.multi_file
@@ -327,110 +163,19 @@ class PipelineDataBlock(metaclass=MetaPipelineDataBlock):
                     return
                 files.append(Path(file_info["location"]))
                 checksums.append(file_info["checksums"])
-
-        # Second step - pass through parser(s)
-        file_folder = CONFIG.FILE_DIRECTORY
-        parser_output_df: "list[pd.DataFrame]" = []
-        parser_checksums: list[str] = []
-
-        for index, file in enumerate(files):
-            # TODO Need to refactor in case we want to parse two .txt(/any other extension) files that are the same..
-            # TODO logic could also do with refining, depending on use case.
-            for parser in self.parser_functions:
-                LOGGER.info("Processing file %s", file)
-                try:
-                    parser_checksum, result = parser.perform_with_cache(
-                        checksums[index], file_folder, file
-                    )
-                except Exception as exc:
-                    warnings.warn(
-                        f"Could not parse file {file} as {self.blocktype} data. Error: {exc}"
-                    )
-                else:
-                    if result is not None:
-                        if type(result) is not list:
-                            result = [result]
-                        parser_output_df.extend(result)
-                        parser_checksums.append(parser_checksum)
-                        break
-            else:
-                LOGGER.warning(
-                    "This file failed to be processed successfully by any of the available parsers"
-                )
-        # Third step - pass through the processors
-        processor_input_name: str = "files"
-        processor_input: list[pd.DataFrame] = parser_output_df
-        processor_output: list[pd.DataFrame] = []
-        processor_checksums: list[str] = parser_checksums
-
-        for ind, processor_group in enumerate(self.processor_functions):
-            processor_output = []
-            output_checksums: list[str] = []
-            # Check the length of input compared to the number of processors available
-            if len(processor_input) != len(processor_group):
-                LOGGER.info(
-                    "Different number of %s to available processors (%s!=%s). "
-                    "Assume lists of dfs should be passed to each processor.",
-                    processor_input_name,
-                    len(processor_input),
-                    len(processor_group),
-                )
-                for processor in processor_group:
-                    process_checksum, result = processor.perform_with_cache(
-                        "".join(processor_checksums), file_folder, processor_input, **self.data
-                    )
-                    if result is not None:
-                        if type(result) is not list:
-                            result = [result]
-                        processor_output.extend(result)
-                        output_checksums.extend([process_checksum] * len(result))
-            else:
-                LOGGER.info(
-                    "Exact match between processors and %s, "
-                    "proceeding to perform one-to-one processing.",
-                    processor_input_name,
-                )
-                for i, processor in enumerate(processor_group):
-                    LOGGER.info(processor)
-                    process_checksum, result = processor.perform_with_cache(
-                        processor_checksums[i], file_folder, processor_input[i], **self.data
-                    )
-                    if result is not None:
-                        if type(result) is not list:
-                            result = [result]
-                        processor_output.extend(result)
-                        output_checksums.extend([process_checksum] * len(result))
-            # Reset variables for next iteration
-            processor_input_name = f"output from layer {ind}"
-            processor_input = processor_output
-            processor_checksums = output_checksums
-
-        # Fourth step - plot
-        if len(processor_output) == 0:
-            LOGGER.info("There is no dataframe to plot, processors are most presumably malformed")
-        elif len(processor_output) > 1 and not self.plotter_function.list_df_input:
-            LOGGER.warning(
-                "There is not one sole dataframe to plot (len = %s) and the plotter does not support multiple dataframe, "
-                "processors are most presumably malformed.",
-                len(processor_output),
-            )
-        elif not self.plotter_function.list_df_input:
-            self.data["bokeh_plot_data"] = self.plotter_function.perform(
-                processor_output[0], **self.data
-            )
-        else:
-            self.data["bokeh_plot_data"] = self.plotter_function.perform(
-                processor_output, **self.data
-            )
+        # Perform pipeline step
+        self.pipeline.perform_entire_pipeline(
+            data=self.data, file_folder=CONFIG.FILE_DIRECTORY, files=files, checksums=checksums
+        )
 
     def to_web(self) -> dict[str, Any]:
         """Returns a JSON serializable dictionary to render the data block on the web."""
         block_errors = []
         block_warnings = []
-        if self.plotter_function or self.processor_functions or self.parser_functions:
+        if self.pipeline.exists():
             with warnings.catch_warnings(record=True) as captured_warnings:
                 try:
-                    self.perform_entire_pipeline()
+                    self.perform_operations()
                 except Exception as e:
                     tb_list = traceback.extract_tb(e.__traceback__)
                     last = tb_list[-1]
@@ -478,9 +223,9 @@ class PipelineDataBlock(metaclass=MetaPipelineDataBlock):
 
         for event in events:
             # Match the event to any registered by the block
-            if (event_name := event.pop("event_name")) in self.event_functions.keys():
+            if (event_name := event.pop("event_name")) in self.pipeline.event_functions.keys():
                 # Bind the method to the instance before calling
-                event_stage = self.event_functions[event_name]
+                event_stage = self.pipeline.event_functions[event_name]
                 try:
                     event_stage.perform(self.data, **event)
                 except Exception as e:
@@ -530,8 +275,9 @@ class PipelineDataBlock(metaclass=MetaPipelineDataBlock):
         self.data.update(self.block_db_model(**data).model_dump(exclude_unset=True))
         return self
 
-    @staticmethod
+    @classmethod
     def define(
+        cls,
         class_name: str,
         *,
         parser: ParserStage | list[ParserStage] | None = None,
@@ -545,18 +291,19 @@ class PipelineDataBlock(metaclass=MetaPipelineDataBlock):
         multi_file: bool = False,
         defaults: dict[str, Any] | None = None,
     ):
-        return MetaPipelineDataBlock(
+        return type(
             class_name,
-            (PipelineDataBlock,),
-            {},
-            parser=parser,
-            processor=processor,
-            plotter=plotter,
-            events=events,
-            blocktype=blocktype,
-            block_name=name,
-            description=description,
-            accepted_file_extensions=accepted_file_extensions,
-            multi_file=multi_file,
-            defaults=defaults,
+            (cls,),
+            {
+                "pipeline": Pipeline(
+                    parser=parser, processor=processor, plotter=plotter, events=events
+                ),
+                "blocktype": blocktype or cls.blocktype,
+                "name": name or cls.name,
+                "description": description or cls.description,
+                "accepted_file_extensions": accepted_file_extensions
+                or cls.accepted_file_extensions,
+                "multi_file": multi_file or cls.multi_file,
+                "defaults": defaults or {},
+            },
         )
